fix(ui): correctly save videos to gallery on mobile
This commit fixes an issue where videos downloaded from attachments were not correctly saved to the device's gallery on mobile platforms.

The `saveFile` method in `StreamAttachmentHandler` now uses `Gal.putVideo` for video files and `Gal.putImage` for image files. If the file type is neither image nor video, it's saved to the app's directory without being copied to the gallery.

Additionally, the temporary file is now only deleted if it was successfully copied to the gallery.

+++ packages/stream_chat_flutter/lib/src/attachment/handler/stream_attachment_handler_io.dart
@@ -198,10 +198,19 @@ class StreamAttachmentHandler extends StreamAttachmentHandlerBase {
 
     // Now that the file is saved, we need to copy it to the user's gallery
     // because the gallery only shows files that are in the gallery folder.
-    await Gal.putImage(path);
+    final copyToGallery = switch (file.mimeType) {
+      final type? when type.startsWith('image/') => Gal.putImage,
+      final type? when type.startsWith('video/') => Gal.putVideo,
+      _ => null,
+    };
 
-    // Once the file is copied to the gallery, we can delete the temporary file.
-    await file.delete();
+    if (copyToGallery != null) {
+      await copyToGallery.call(path);
+
+      // If the file was successfully copied to the gallery, we can safely
+      // delete the temporary file.
+      await file.delete();
+    }
 
     return path;
   }
